backend/fix: commonize ride receipt line items
